requirements.txt
evdev==1.6.0
MouseInfo==0.1.3
obs-websocket-py==0.5.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-xlib==0.31
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
websocket-client==1.3.3
websockets==10.3
